chore: move `util.core` and `plugin` to the `core` namespace
