+++ js/entries/plugins/foundation.core.js
@@ -6,7 +6,7 @@ Foundation.addToJquery($);
 // These are now separated out, but historically were a part of this module,
 // and since this is here for backwards compatibility we include them in
 // this entry.
-import {rtl, GetYoDigits, transitionend, RegExpEscape} from '../../foundation.util.core';
+import { rtl, GetYoDigits, transitionend, RegExpEscape } from '../../foundation.core.utils';
 Foundation.rtl = rtl;
 Foundation.GetYoDigits = GetYoDigits;
 Foundation.transitionend = transitionend;
@@ -14,8 +14,7 @@ Foundation.RegExpEscape = RegExpEscape;
 
 // Every plugin depends on plugin now, we can include that on the core for the
 // script inclusion path.
-
-import { Plugin } from '../../foundation.plugin';
+import { Plugin } from '../../foundation.core.plugin';
 Foundation.Plugin = Plugin;
