+++ js/foundation.orbit.js
@@ -5,8 +5,8 @@ import { Keyboard } from './foundation.util.keyboard';
 import { Motion } from './foundation.util.motion';
 import { Timer } from './foundation.util.timer';
 import { onImagesLoaded } from './foundation.util.imageLoader';
-import { GetYoDigits } from './foundation.util.core';
-import { Plugin } from './foundation.plugin';
+import { GetYoDigits } from './foundation.core.utils';
+import { Plugin } from './foundation.core.plugin';
 import { Touch } from './foundation.util.touch'
