+++ js/foundation.responsiveAccordionTabs.js
@@ -2,8 +2,8 @@
 
 import $ from 'jquery';
 import { MediaQuery } from './foundation.util.mediaQuery';
-import { GetYoDigits } from './foundation.util.core';
-import { Plugin }from './foundation.plugin';
+import { GetYoDigits } from './foundation.core.utils';
+import { Plugin }from './foundation.core.plugin';
 
 import { Accordion } from './foundation.accordion';
 import { Tabs } from './foundation.tabs';
