+++ js/foundation.util.motion.js
@@ -1,7 +1,7 @@
 'use strict';
 
 import $ from 'jquery';
-import { transitionend } from './foundation.util.core';
+import { transitionend } from './foundation.core.utils';
 
 /**
  * Motion module.
